style: HTML conventions to view files V2 - Standadize ID and class naimg conventions across HTML templates - consider Mobile resposive, js moduleize
Related to: #51
