+++ view/src/main/resources/static/js/chat/chatList/newChatButton.js
