+++ view/src/main/resources/static/js/chat/chatList/startChat.js
