+++ view/src/main/resources/static/js/chat/chatList/startChatButton.js
