+++ view/src/main/resources/static/js/chat/chatRoom/index.js
@@ -1,0 +1,9 @@
+import { initWebSocket } from './websocketConnection.js';
+import { initializeChat } from './messageHandling.js';
+import { setupEventListeners } from './uiUpdates.js';
+
+document.addEventListener('DOMContentLoaded', () => {
+    initWebSocket();
+    initializeChat();
+    setupEventListeners();
+});
