+++ view/src/main/resources/templates/chat/chatroom.html
@@ -60,6 +60,7 @@
     var initialMessages = /*[[${messages}]]*/ [];
     /*]]>*/
 </script>
-<script type="module" src="/js/chat/chatroom.js"></script>
+<!--<script type="module" src="/js/chat/chatroom.js"></script>-->
+<script type="module" src="/js/chat/chatRoom/index.js"></script>
 </body>
 </html>
