+++ view/src/main/resources/templates/chat/messages.html
@@ -60,6 +60,7 @@
     </div>
 </template>
 
-<script type="module" src="/js/chat/index.js"></script>
+<!--<script type="module" src="/js/chat/index.js"></script>-->
+<script type="module" src="/js/chat/chatList/index.js"></script>
 </body>
 </html>
